Update SVCCCD.R
Certain imbalance and overlap levels caused decision values to all be same value which affected everything downstream. Added buffer to account for this, and modified SVM to account for both cost and class weight to avoid cases of single imbalanced class not being able to be separated easily. Byproduct is a better overall classifier.